Shared domain state
Each domain offers shared methods that other domains can use. With
the advent of cached prepare statements, we also want to take
advantage of the ability to reuse those statements without recreating
them each transaction.

The concept is simple, create a shared state that can be composed
together with other shared states (users + permissions) and allow
others to consume them.
